requirements.txt
Django==1.6.2
South==0.7.5
djangorestframework==2.4.5